Installer allows running on existing environments

+++ install.sh
@@ -1,18 +1,28 @@
 #!/bin/sh
 
-zshrc() {
-    echo "cloning zsh-autosuggestions."
-    git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-    echo "cloning zsh-syntax-highlighting."
-    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
-    echo "cloning powerlevel10k."
-    git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
-    echo "import zshrc."
-    cat .zshrc > $HOME/.zshrc
+get_zsh_plugin() {
+    install_dir=$1
+    repo=$2
+    if [[ -d $install_dir ]]; then
+        echo "Skipping Install - directory exists: ${install_dir}"
+    else
+        echo "cloning $repo into $install_dir."
+        git clone $repo $install_dir
+    fi
 }
 
-zshrc
+zsh_custom=${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}
+get_zsh_plugin ${zsh_custom}/plugins/zsh-autosuggestions https://github.com/zsh-users/zsh-autosuggestions.git
+get_zsh_plugin ${zsh_custom}/plugins/zsh-syntax-highlighting https://github.com/zsh-users/zsh-syntax-highlighting.git
+get_zsh_plugin ${zsh_custom}/themes/powerlevel10k "--depth=1 https://github.com/romkatv/powerlevel10k.git"
+
+if [[ -f $HOME/.zshrc ]]; then
+    echo "zshrc already exists, not updating"
+else
+    echo "import zshrc."
+    cat .zshrc >$HOME/.zshrc
+fi
 
 # Pre-Commit Install and configure
 echo "Installing Pre-Commit."
-pip install pre-commit
+pip3 install --user pre-commit
